Added a utility function to 'sys/yaos.h' and a command to the PCI module.

+++ modules/pci/arch/x86.c
@@ -2,7 +2,7 @@
 
 #include <kernel/arch/x86/pio.h>
 
-UINT32 ReadPCIDeviceConfig32(PCI_DEVICE Device, UINT8 Offset)
+UINT32 ReadPCIDeviceConfig32(PCI_DEVICE Device, UINT Offset)
 {
   UINT32 Address = ((UINT32)1 << 31) | ((UINT32)Device << 8) | (Offset & 0xFC);
   WritePort32(0xCF8, Address);
